Fix API tests after rebase
