+++ api-tester/src/defaults/kotlin/com/revenuecat/apitester/kotlin/PurchasesAPI.kt
@@ -1,13 +1,10 @@
 package com.revenuecat.apitester.kotlin
 
-import android.app.Activity
 import android.content.Context
 import com.revenuecat.purchases.CacheFetchPolicy
 import com.revenuecat.purchases.CustomerInfo
 import com.revenuecat.purchases.EntitlementVerificationMode
 import com.revenuecat.purchases.ExperimentalPreviewRevenueCatPurchasesAPI
-import com.revenuecat.purchases.Package
-import com.revenuecat.purchases.PurchaseParams
 import com.revenuecat.purchases.Purchases
 import com.revenuecat.purchases.PurchasesConfiguration
 import com.revenuecat.purchases.PurchasesError
@@ -15,16 +12,11 @@ import com.revenuecat.purchases.Store
 import com.revenuecat.purchases.awaitCustomerInfo
 import com.revenuecat.purchases.getCustomerInfoWith
 import com.revenuecat.purchases.interfaces.LogInCallback
-import com.revenuecat.purchases.interfaces.PurchaseCallback
 import com.revenuecat.purchases.interfaces.ReceiveCustomerInfoCallback
 import com.revenuecat.purchases.interfaces.SyncPurchasesCallback
 import com.revenuecat.purchases.logInWith
 import com.revenuecat.purchases.logOutWith
 import com.revenuecat.purchases.models.BillingFeature
-import com.revenuecat.purchases.models.GoogleProrationMode
-import com.revenuecat.purchases.models.StoreProduct
-import com.revenuecat.purchases.models.StoreTransaction
-import com.revenuecat.purchases.models.SubscriptionOption
 import com.revenuecat.purchases.restorePurchasesWith
 import com.revenuecat.purchases.syncPurchasesWith
 import java.util.concurrent.ExecutorService
@@ -71,39 +63,6 @@ private class PurchasesAPI {
         purchases.onAppForegrounded()
 
         val store: Store = purchases.store
-    }
-
-    @SuppressWarnings("LongParameterList", "EmptyFunctionBlock")
-    fun checkPurchasing(
-        purchases: Purchases,
-        activity: Activity,
-        storeProduct: StoreProduct,
-        packageToPurchase: Package,
-        subscriptionOption: SubscriptionOption,
-    ) {
-        val purchaseCallback = object : PurchaseCallback {
-            override fun onCompleted(storeTransaction: StoreTransaction, customerInfo: CustomerInfo) {}
-            override fun onError(error: PurchasesError, userCancelled: Boolean) {}
-        }
-
-        val oldProductId = "old"
-        val prorationMode = GoogleProrationMode.IMMEDIATE_WITH_TIME_PRORATION
-        val isPersonalizedPrice = true
-
-        val purchasePackageBuilder: PurchaseParams.Builder = PurchaseParams.Builder(activity, packageToPurchase)
-        purchasePackageBuilder.oldProductId(oldProductId).googleProrationMode(prorationMode)
-        val purchasePackageParams: PurchaseParams = purchasePackageBuilder.build()
-        purchases.purchase(purchasePackageParams, purchaseCallback)
-
-        val purchaseProductBuilder: PurchaseParams.Builder = PurchaseParams.Builder(activity, storeProduct)
-        purchaseProductBuilder.oldProductId(oldProductId).googleProrationMode(prorationMode)
-        val purchaseProductParams: PurchaseParams = purchaseProductBuilder.build()
-        purchases.purchase(purchaseProductParams, purchaseCallback)
-
-        val purchaseOptionBuilder: PurchaseParams.Builder = PurchaseParams.Builder(activity, subscriptionOption)
-        purchaseOptionBuilder.oldProductId(oldProductId).googleProrationMode(prorationMode)
-        val purchaseOptionsParams: PurchaseParams = purchaseOptionBuilder.build()
-        purchases.purchase(purchaseOptionsParams, purchaseCallback)
     }
 
     @Suppress("LongMethod", "LongParameterList")
