+++ api-tester/src/main/java/com/revenuecat/apitester/kotlin/PurchasesCommonAPI.kt
@@ -3,7 +3,6 @@ package com.revenuecat.apitester.kotlin
 import android.app.Activity
 import android.content.Context
 import com.revenuecat.purchases.CustomerInfo
-import com.revenuecat.purchases.EntitlementVerificationMode
 import com.revenuecat.purchases.ExperimentalPreviewRevenueCatPurchasesAPI
 import com.revenuecat.purchases.LogHandler
 import com.revenuecat.purchases.LogLevel
@@ -12,7 +11,6 @@ import com.revenuecat.purchases.Package
 import com.revenuecat.purchases.ProductType
 import com.revenuecat.purchases.PurchaseParams
 import com.revenuecat.purchases.Purchases
-import com.revenuecat.purchases.PurchasesConfiguration
 import com.revenuecat.purchases.PurchasesError
 import com.revenuecat.purchases.awaitOfferings
 import com.revenuecat.purchases.getOfferingsWith
@@ -28,10 +26,9 @@ import com.revenuecat.purchases.models.StoreTransaction
 import com.revenuecat.purchases.models.SubscriptionOption
 import com.revenuecat.purchases.purchaseWith
 import java.net.URL
-import java.util.concurrent.ExecutorService
 
 @OptIn(ExperimentalPreviewRevenueCatPurchasesAPI::class)
-@Suppress("unused", "UNUSED_VARIABLE", "EmptyFunctionBlock", "DEPRECATION")
+@Suppress("unused", "UNUSED_VARIABLE", "EmptyFunctionBlock")
 private class PurchasesCommonAPI {
     @SuppressWarnings("LongParameterList")
     fun check(
@@ -131,19 +128,9 @@ private class PurchasesCommonAPI {
     }
 
     @Suppress("ForbiddenComment")
-    fun checkConfiguration(context: Context, executorService: ExecutorService) {
+    fun checkConfiguration(context: Context) {
         val features: List<BillingFeature> = ArrayList()
         val configured: Boolean = Purchases.isConfigured
-
-        val build = PurchasesConfiguration.Builder(context, apiKey = "", appUserID = "")
-            .observerMode(true)
-            .observerMode(false)
-            .service(executorService)
-            .diagnosticsEnabled(true)
-            .entitlementVerificationMode(EntitlementVerificationMode.INFORMATIONAL)
-            .build()
-
-        Purchases.configure(build)
 
         Purchases.canMakePayments(context, features) { _: Boolean -> }
         Purchases.canMakePayments(context) { _: Boolean -> }
